issue: 3788369 Rename thread_local_event_handler
It's not thread_local specific anymore. This event handler will also be
used within poll_group to provide TCP timers for the respective sockets.
