+++ src/core/sock/sock-redirect.cpp
@@ -49,7 +49,7 @@
 #include "utils/lock_wrapper.h"
 #include <proto/ip_frag.h>
 #include <dev/buffer_pool.h>
-#include <event/thread_local_event_handler.h>
+#include <event/event_handler_manager_local.h>
 #include <event/vlogger_timer_handler.h>
 #include <iomux/poll_call.h>
 #include <iomux/select_call.h>
@@ -710,7 +710,7 @@ inline int epoll_wait_helper(int __epfd, struct epoll_event *__events, int __max
     }
 
     if (safe_mce_sys().tcp_ctl_thread == option_tcp_ctl_thread::CTL_THREAD_DELEGATE_TCP_TIMERS) {
-        g_thread_local_event_handler.do_tasks();
+        g_event_handler_manager_local.do_tasks();
     }
 
     epoll_event extra_events_buffer[__maxevents];
